Move precedence list out of Node.js
